fix codecheck

+++ frameworks/js/napi/src/socket/napi_bluetooth_spp_client.cpp
@@ -244,10 +244,10 @@ napi_value NapiSppClient::SppWrite(napi_env env, napi_callback_info info)
     NAPI_BT_ASSERT_RETURN_FALSE(env, clientMap[id] > 0, BT_ERR_INTERNAL_ERROR);
     OutputStream outputStream = clientMap[id]->client_->GetOutputStream();
     while (totalSize) {
-        unsigned int result = outputStream.Write(totalBuf, totalSize);
+        int result = outputStream.Write(totalBuf, totalSize);
         NAPI_BT_ASSERT_RETURN_FALSE(env, result > 0, BT_ERR_SPP_IO);
-        totalSize = totalSize - result;
-        totalBuf += result;
+        totalSize = totalSize - static_cast<size_t>(result);
+        totalBuf += static_cast<size_t>(result);
         isOK = true;
     }
     return NapiGetBooleanRet(env, isOK);
